+++ src/main/java/blackjack/view/OutputView.java
@@ -1,10 +1,10 @@
 package blackjack.view;
 
-import blackjack.domain.Dealer;
 import blackjack.domain.Judgement;
-import blackjack.domain.Participant;
-import blackjack.domain.Player;
 import blackjack.domain.card.Card;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participant;
+import blackjack.domain.participant.Player;
 import blackjack.domain.winResult;
 import java.util.List;
 import java.util.Map;
