+++ src/test/java/blackjack/domain/winResultTest.java
@@ -19,6 +19,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.card.Denomination;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Player;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.List;
